refactor: simplify explorer env handling
